backend/emails/lambda/requirements.txt
boto3
python-dotenv
sentence-transformers